requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.3.3
psycopg2-binary==2.9.9